pong update

+++ src/client/pong/pongcom.h
@@ -10,8 +10,8 @@ typedef struct {
 
 typedef struct {
     PLAYER player[2];
-    int ball_x;
-    int ball_y;
+    float ball_x;
+    float ball_y;
 }PONGPacket;
 
 extern PONGPacket pong;
